OT146-57 | minor fix. refactor
