+++ src/main/java/com/alkemy/ong/domain/utils/CustomMultipartFile.java
@@ -1,4 +1,4 @@
-package com.alkemy.ong.domain.cloud;
+package com.alkemy.ong.domain.utils;
 
 import org.springframework.web.multipart.MultipartFile;
